Fix ministamp and minibundle usage for layouts
This fixes the bug causing thrown exception when ministamp tag or
minibundle block is called multiple times with same asset source as
argument. Because this is very common with layouts, it must be allowed.

Also, this removes the restriction of disallowing the use of static
files as asset sources: you can now use Jekyll's static files as asset
sources for Minibundle. Even though doing this is silly (resulting in
duplicated assets in the output directory), the restriction was
artificial now that a proper mechanism was found for the bug mentioned
above.